Move CMomentumRocket to a separate file
