+++ mp/src/game/client/client_momentum.vpc
@@ -294,6 +294,8 @@ $Project "Client (Momentum)"
             $File   "$SRCDIR\game\shared\momentum\mom_gamemovement.h"
             $File   "$SRCDIR\game\shared\momentum\mom_player_shared.h"
             $File   "$SRCDIR\game\shared\momentum\mom_player_shared.cpp"
+            $File   "$SRCDIR\game\shared\momentum\mom_rocket_projectile.h"
+            $File   "$SRCDIR\game\shared\momentum\mom_rocket_projectile.cpp"
             $File   "$SRCDIR\game\shared\momentum\mom_gamerules.cpp"
             $File   "$SRCDIR\game\shared\momentum\mom_gamerules.h"
             $File   "$SRCDIR\game\shared\momentum\mom_usermessages.cpp"
