+++ mp/src/game/server/server_momentum.vpc
@@ -145,6 +145,8 @@ $Project "Server (Momentum)"
             $File   "$SRCDIR\game\shared\momentum\mom_gamerules.h"
             $File   "$SRCDIR\game\shared\momentum\mom_player_shared.h"
             $File   "$SRCDIR\game\shared\momentum\mom_player_shared.cpp"
+            $File   "$SRCDIR\game\shared\momentum\mom_rocket_projectile.h"
+            $File   "$SRCDIR\game\shared\momentum\mom_rocket_projectile.cpp"
             $File   "$SRCDIR\game\shared\momentum\mom_usermessages.cpp"
             $File   "$SRCDIR\game\shared\momentum\mom_shareddefs.h"
             $File   "$SRCDIR\game\shared\momentum\mom_modulecomms.cpp"
